fix(two-plane): authenticate the relayed pairing exchange and isolate its test
The relayed pairing request went out unauthenticated. submitConnectPairing took
`fetchImpl: typeof fetch = fetch`, and a default parameter binds the global as
it was when the module was evaluated — the unwrapped original, not the wrapper
installApiAuthFetch puts on window.fetch. The relay needs the machine-session
headers that wrapper attaches, so the hub refused the exchange. Resolved at
call time now.

The transport also moves to its own module. One file exported both a transport
function and a component, which react-refresh/only-export-components flags for
good reason; the previous shape carried an eslint-disable instead. The two have
no reason to share a file: the transport is testable without React and the form
has no logic beyond calling it.

tests/connect-pairing.test.ts passed alone and failed in the full GUI run. App
calls installApiAuthFetch() at module scope, so it runs on first import only; a
later test importing App gets the cached module and no install, leaving the
wrapper bound to whichever window imported it first. The test now binds the
wrapper to its own window before mounting, and claude-toggle-race.test.tsx
clears the install latch in afterEach alongside the window it closes. Both are
test isolation rather than product behavior.
